chore: moved `contract` into `program`
