--- packages/contract/.gitignore
@@ -1,3 +1,0 @@
-# Forc
-**/out/debug/**
-**/out/release/**
