--- packages/contract/README.md
@@ -1,51 +1,0 @@
-# `@fuel-ts/contract`
-
-**@fuel-ts/contract** is a sub-module for interacting with **Fuel**.
-
-This module allows for a simple way to serialize calls and transactions to an on-chain contract and deserialize their results and emitted logs.
-
-# Table of contents
-
-- [Documentation](#documentation)
-- [Usage](#usage)
-  - [Installation](#installation)
-  - [Full SDK Installation](#full-sdk-installation)
-- [Contributing](#contributing)
-- [Changelog](#changelog)
-- [License](#license)
-
-## Documentation
-
-See [Fuels-ts Documentation](https://docs.fuel.network/docs/fuels-ts/contracts/)
-
-## Usage
-
-### Installation
-
-```sh
-pnpm add @fuel-ts/contract
-# or
-npm add @fuel-ts/contract
-```
-
-### Full SDK Installation
-
-Alternatively, we recommend you install the [complete SDK](https://github.com/FuelLabs/fuels-ts) using the umbrella package:
-
-```sh
-pnpm add fuels
-# or
-npm add fuels
-```
-
-## Contributing
-
-In order to contribute to `@fuel-ts/contract`, please see the main [fuels-ts](https://github.com/FuelLabs/fuels-ts) monorepo.
-
-## Changelog
-
-The `@fuel-ts/contract` changelog can be found at [CHANGELOG](./CHANGELOG.md).
-
-## License
-
-The primary license for `@fuel-ts/contract` is `Apache 2.0`, see [LICENSE](./LICENSE).
