--- packages/contract/package.json
@@ -1,60 +1,0 @@
-{
-  "name": "@fuel-ts/contract",
-  "version": "0.101.1",
-  "description": "",
-  "author": "Fuel Labs <contact@fuel.sh> (https://fuel.network/)",
-  "main": "dist/index.js",
-  "module": "dist/index.mjs",
-  "types": "dist/index.d.ts",
-  "engines": {
-    "node": "^20.0.0 || ^22.0.0 || ^24.0.0"
-  },
-  "exports": {
-    ".": {
-      "types": "./dist/index.d.ts",
-      "require": "./dist/index.js",
-      "import": "./dist/index.mjs"
-    },
-    "./test-utils": {
-      "types": "./dist/test-utils.d.ts",
-      "require": "./dist/test-utils.js",
-      "import": "./dist/test-utils.mjs"
-    }
-  },
-  "typesVersions": {
-    "*": {
-      "test-utils": [
-        "./dist/test-utils.d.ts"
-      ]
-    }
-  },
-  "files": [
-    "dist"
-  ],
-  "scripts": {
-    "pretest": "pnpm build:forc",
-    "build:forc": "pnpm fuels-forc build -p ./test/fixtures/forc-projects --release",
-    "build": "tsup",
-    "postbuild": "tsx ../../scripts/postbuild.ts"
-  },
-  "license": "Apache-2.0",
-  "dependencies": {
-    "@fuel-ts/abi-coder": "workspace:*",
-    "@fuel-ts/account": "workspace:*",
-    "@fuel-ts/crypto": "workspace:*",
-    "@fuel-ts/errors": "workspace:*",
-    "@fuel-ts/hasher": "workspace:*",
-    "@fuel-ts/math": "workspace:*",
-    "@fuel-ts/merkle": "workspace:*",
-    "@fuel-ts/program": "workspace:*",
-    "@fuel-ts/transactions": "workspace:*",
-    "@fuel-ts/utils": "workspace:*",
-    "@fuels/vm-asm": "0.60.2",
-    "ramda": "0.30.1"
-  },
-  "devDependencies": {
-    "@fuel-ts/utils": "workspace:*",
-    "@internal/forc": "workspace:*",
-    "@types/ramda": "0.30.2"
-  }
-}
