--- packages/contract/src/index.ts
@@ -1,6 +1,0 @@
-export {
-  default as ContractFactory,
-  DeployContractOptions,
-  DeployContractResult,
-} from './contract-factory';
-export { getContractRoot, getContractStorageRoot, getContractId, hexlifyWithPrefix } from './util';
