--- packages/contract/test/fixtures/forc-projects/Forc.toml
@@ -1,2 +1,0 @@
-[workspace]
-members = ["simple-contract"]
