--- packages/contract/tsup.config.ts
@@ -1,12 +1,0 @@
-import { tsupDefaults } from '@internal/tsup';
-import type { Options } from 'tsup';
-
-const configs: Options = {
-  ...tsupDefaults,
-  entry: {
-    index: 'src/index.ts',
-    'test-utils': 'src/test-utils.ts',
-  },
-};
-
-export default configs;
