--- packages/contract/typedoc.json
@@ -1,6 +1,0 @@
-{
-  "$schema": "https://typedoc.org/schema.json",
-  "extends": ["../../typedoc.base.json"],
-  "entryPoints": ["src/index.ts", "src/test-utils.ts"],
-  "readme": "none"
-}
