+++ packages/fuels/package.json
@@ -67,7 +67,6 @@
     "@fuel-ts/abi-typegen": "workspace:*",
     "@fuel-ts/account": "workspace:*",
     "@fuel-ts/address": "workspace:*",
-    "@fuel-ts/contract": "workspace:*",
     "@fuel-ts/crypto": "workspace:*",
     "@fuel-ts/errors": "workspace:*",
     "@fuel-ts/hasher": "workspace:*",
