+++ packages/fuels/src/test-utils.ts
@@ -1,4 +1,4 @@
-export * from '@fuel-ts/contract/test-utils';
 export * from '@fuel-ts/account/test-utils';
 export * from '@fuel-ts/errors/test-utils';
+export * from '@fuel-ts/program/test-utils';
 export * from '@fuel-ts/utils/test-utils';
