+++ packages/program/src/index.ts
@@ -1,3 +1,4 @@
+// Program
 export * from './types';
 export * from './utils';
 export { FunctionInvocationScope } from './functions/invocation-scope';
@@ -6,4 +7,19 @@ export { Contract } from './contract';
 export { ScriptRequest } from './script-request';
 export { InstructionSet } from './instruction-set';
 export * from './response';
+
+// Script
 export { Script } from './script';
+
+// Contract
+export {
+  default as ContractFactory,
+  DeployContractOptions,
+  DeployContractResult,
+} from './contract/contract-factory';
+export {
+  getContractRoot,
+  getContractStorageRoot,
+  getContractId,
+  hexlifyWithPrefix,
+} from './contract/utils';
