+++ packages/program/test/fixtures/forc-projects/simple-contract/src/main.sw
