+++ packages/program/tsup.config.ts
@@ -1,3 +1,12 @@
-import { index } from '@internal/tsup';
+import { tsupDefaults } from '@internal/tsup';
+import type { Options } from 'tsup';
 
-export default index;
+const configs: Options = {
+  ...tsupDefaults,
+  entry: {
+    index: 'src/index.ts',
+    'test-utils': 'src/test-utils.ts',
+  },
+};
+
+export default configs;
