+++ packages/program/typedoc.json
@@ -1,6 +1,6 @@
 {
   "$schema": "https://typedoc.org/schema.json",
   "extends": ["../../typedoc.base.json"],
-  "entryPoints": ["src/index.ts"],
+  "entryPoints": ["src/index.ts", "src/test-utils.ts"],
   "readme": "none"
 }
