+++ packages/recipes/scripts/build-recipes.ts
@@ -7,8 +7,8 @@ execSync(`fuels-typegen -i src/contracts/**/*-abi.json -o src/types`);
 const supportedRecipes = ['Src14OwnedProxy'].map((s) => [s, `${s}Factory`]).flat();
 const importReplacementMap = {
   'Contract as __Contract': '@fuel-ts/program',
-  'ContractFactory as __ContractFactory': '@fuel-ts/contract',
-  DeployContractOptions: '@fuel-ts/contract',
+  'ContractFactory as __ContractFactory': '@fuel-ts/program',
+  DeployContractOptions: '@fuel-ts/program',
   Interface: '@fuel-ts/abi-coder',
   Provider: '@fuel-ts/account',
   Account: '@fuel-ts/account',
